refactor: extract commands into cogs and config module
